gk-rpc-common/src/main/java/com/smile/gkrpc/common/utils/ReflectionUtils.java
package com.smile.gkrpc.common.utils;

import java.lang.reflect.InvocationTargetException;
import java.lang.reflect.Method;
import java.lang.reflect.Modifier;
import java.util.ArrayList;
import java.util.List;

/**
 * 反射工具类
 */
public class ReflectionUtils {

    /**
     * 根据class创建对线
     *
     * @param clazz 待创建对象的类
     * @param <T>   对线类型
     * @return 创建好的对象
     */
    public static <T> T newInstance(Class<T> clazz) {
        try {
            return clazz.newInstance();
        } catch (InstantiationException | IllegalAccessException e) {
            throw new IllegalStateException(e);
        }
    }

    /**
     * 获取某个class的公共方法
     *
     * @param clazz 类
     * @return 返回当前类的公共方法
     */
    public static Method[] getPublicMethods(Class clazz) {
        Method[] declaredMethods = clazz.getDeclaredMethods();
        List<Method> methods = new ArrayList<>();
        for (Method declaredMethod : declaredMethods) {
            if (Modifier.isPublic(declaredMethod.getModifiers())) {
                methods.add(declaredMethod);
            }
        }
        return methods.toArray(new Method[0]);
    }


    /**
     * 调用指定对象的指定方法
     * @param object 被调用方法的对象
     * @param method 被调用方法
     * @param args 参数
     * @return 返回结果
     */
    public static Object invoke(Object object, Method method, Object... args) {
        try {
            return method.invoke(object, args);
        } catch (IllegalAccessException | InvocationTargetException e) {
            throw new IllegalStateException(e);
        }
    }
}
